pyproject.toml
[tool.poetry]
name = "habit-manager"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-cors-headers = "^4.6.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
attrs = "^24.3.0"
requests = "^2.32.3"
celery = "^5.4.0"
billiard = "4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
